+++ NEWS.md
@@ -1,5 +1,7 @@
 # vroom (development version)
 
+* Subsetting datetime columns now works with `NA` indices (#236).
+
 # vroom 1.2.1
 
 * `vroom()` now writes the column names if given an input with no rows (#213)